Read number of steps when opening watchface.
Previously number of steps was only updated once there was an event sent with num steps updated.

+++ app/src/accelerometer.h
@@ -56,7 +56,7 @@ int accelerometer_fetch_xyz(int16_t *x, int16_t *y, int16_t *z);
 
 int accelerometer_fetch_temperature(struct sensor_value *temperature);
 
-int accelerometer_fetch_num_steps(int16_t *num_steps);
+int accelerometer_fetch_num_steps(uint32_t *num_steps);
 
 int accelerometer_reset_step_count(void);
